+++ app/Policies/UserPolicy.php
@@ -201,7 +201,7 @@ class UserPolicy
             return true;
         }
 
-        $membershipDuration = (int) now()->diffInDays($user->created_at ?? now(), true);
+        $membershipDuration = now()->diffInDays($user->created_at ?? now(), true);
         if ($membershipDuration >= 14) {
             return true;
         }
